Improve RgbLedSequencer sleep status
The RGB LED Sequencer will now **only** wake up on play sequence and continue commands, and hardware button interrupts.
If the RgbLedSequencer is asleep and a command needing to use the 25LC1024 is used, the 25LC1024 will be awoken temporarily (the low power status and disabled state of other components is unaffected) for these commands, then put back to sleep.
Fixed an issue where the TLC5940 would flicker when awake and turn the LEDs on without the 7 segment display if the interrupt button wasn't pressed long enough to detect the interrupt type.